split data dump and fix pylint errors
